Move collapse_interchangable_instances to src/librustc_mir
