+++ src/librustc_mir/lib.rs
@@ -31,6 +31,7 @@ Rust MIR: a lowered representation of Rust. Also: an experiment!
 #![cfg_attr(stage0, feature(inclusive_range_syntax))]
 #![feature(macro_vis_matcher)]
 #![cfg_attr(stage0, feature(match_default_bindings))]
+#![feature(macro_lifetime_matcher)]
 #![feature(exhaustive_patterns)]
 #![feature(range_contains)]
 #![feature(rustc_diagnostic_macros)]
@@ -82,6 +83,7 @@ pub fn provide(providers: &mut Providers) {
     transform::provide(providers);
     providers.const_eval = interpret::const_eval_provider;
     providers.check_match = hair::pattern::check_match;
+    providers.collapse_interchangable_instances = monomorphize::deduplicate_instances::collapse_interchangable_instances;
 }
 
 __build_diagnostic_array! { librustc_mir, DIAGNOSTICS }
